Use new storybook structure
